+++ buildroot-external/package/libxmlrpcxx/src/XmlRpcServer.cpp
@@ -132,7 +132,7 @@ XmlRpcServer::bindAndListen(const char* ip, int port, int backlog /*= 5*/)
 
 // Process client requests for the specified time
 void 
-XmlRpcServer::work(long msTime)
+XmlRpcServer::work(int32_t msTime)
 {
 //  XmlRpcUtil::log(2, "XmlRpcServer::work: waiting for a connection");
   _disp.work(msTime);
